feat: implement optimized date formatting using Intl.DateTimeFormat for improved performance in dashboard data processing
